Use parseFloat() instead of parseInt() for time
This will allow people to input things like 3.5h. I think the main
use case is hour input.

+++ src/background.js
@@ -14,7 +14,7 @@ resetTimers();
 loadAudios();
 
 function parseTime(str) {
-  var num = parseInt(str);
+  var num = parseFloat(str);
   if (isNaN(num))
     return null;
   var last = str.charAt(str.length - 1);
